+++ app/src/main/java/kr/co/nottodo/view/monthcalendar/NoChangeAnimationRecyclerView.kt
@@ -1,0 +1,18 @@
+package kr.co.nottodo.view.monthcalendar
+
+import android.content.Context
+import android.util.AttributeSet
+import androidx.recyclerview.widget.RecyclerView
+import androidx.recyclerview.widget.SimpleItemAnimator
+
+class NoChangeAnimationRecyclerView @JvmOverloads constructor(
+    context: Context,
+    attrs: AttributeSet? = null,
+    defStyleAttr: Int = 0
+) : RecyclerView(context, attrs, defStyleAttr) {
+
+    init {
+        (itemAnimator as SimpleItemAnimator).supportsChangeAnimations = false
+    }
+
+}
