+++ app/src/main/java/kr/co/nottodo/view/monthcalendar/NoRippleRecyclerView.kt
@@ -1,0 +1,16 @@
+package kr.co.nottodo.view.monthcalendar
+
+import android.content.Context
+import android.util.AttributeSet
+import androidx.recyclerview.widget.RecyclerView
+
+class NoRippleRecyclerView @JvmOverloads constructor(
+    context: Context,
+    attrs: AttributeSet? = null,
+    defStyle: Int = 0
+) : RecyclerView(context, attrs, defStyle) {
+
+    init {
+        overScrollMode = OVER_SCROLL_NEVER
+    }
+}
