+++ app/src/main/java/kr/co/nottodo/view/monthcalendar/util/CalendarUtil.kt
@@ -32,9 +32,10 @@ fun Calendar.isBefore(otherCalendar: Calendar): Boolean {
 }
 
 fun Calendar.isBeforeCalendar(otherCalendar: Calendar): Boolean {
-    return get(YEAR) == otherCalendar.get(YEAR)
-            && get(MONTH) == otherCalendar.get(MONTH)
-            && get(DAY_OF_MONTH) < otherCalendar.get(DAY_OF_MONTH)
+    if (get(YEAR) < otherCalendar.get(YEAR)) return true
+    if (get(MONTH) < otherCalendar.get(MONTH)) return true
+    if (get(DAY_OF_MONTH) <= otherCalendar.get(DAY_OF_MONTH)) return true
+    return false
 }
 
 // 현재 날짜로 부터 이후의 날인지 체크하는 함수
